+++ leonid_andrushchenko/src/main/java/lesson8/Subject.java
@@ -1,5 +1,14 @@
 package lesson8;
 
-public interface Subject {
-    String getName();
+public class Subject
+{
+    private String name;
+
+    public Subject(String name){
+        this.name = name;
+    }
+
+    public String getName(){
+        return name;
+    }
 }
